Extract RuntimeImageHeapList.
